scripts/requirements.txt
feedparser>=6.0.10
requests>=2.31.0
pyyaml>=6.0.1
youtube-transcript-api>=0.6.2
bilibili-api-python>=16.2.0
python-dateutil>=2.8.2

